+++ src/components/ProductFilter.ts
@@ -1,5 +1,5 @@
 import { ObjectForFilter, ObjectInterface } from '../types/products';
-import { transformParamsToUrl, transformUrlToParams, updateUrl } from '../common/urlHelpers';
+import { updateUrl } from '../router/router';
 import { routes } from '../common/constans';
 
 type Elements = {
@@ -80,12 +80,10 @@ class ProductFilter {
       } else {
         this.currentFilters = this.currentFilters.filter((val) => val !== checkbox.value);
       }
-
-      // const currentUrl = new URL(window.location.href).pathname;
-      // console.log(currentUrl);
-
       const params = getParamsToUrl(this.filterName, this.currentFilters);
       updateUrl(routes.mainSearch, params);
+
+
     });
   }
 }
