+++ src/index.ts
@@ -2,6 +2,6 @@ import './index.html';
 import 'reset-css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './global.scss';
-import initRouter from './rout';
+import initRouter from './router/router';
 
 initRouter();
